fix:rollback useAiChat hook
