+++ apps/masterbots.ai/components/routes/chat/chat.tsx
@@ -1,15 +1,26 @@
 'use client'
 
-import React, { useRef, useEffect } from 'react'
+import { Message, useChat } from 'ai/react'
+import { useScroll } from 'framer-motion'
 import { ChatList } from '@/components/routes/chat/chat-list'
 import { ChatPanel } from '@/components/routes/chat/chat-panel'
 import { ChatScrollAnchor } from '@/components/routes/chat/chat-scroll-anchor'
-import { cn, scrollToBottomOfElement } from '@/lib/utils'
+import { cn, extractBetweenMarkers, scrollToBottomOfElement } from '@/lib/utils'
+import { useAtBottom } from '@/lib/hooks/use-at-bottom'
+import { createThread, getThread, saveNewMessage } from '@/services/hasura'
+import { ChatRequestOptions, CreateMessage } from 'ai'
+import { uniqBy } from 'lodash'
+import { Chatbot } from 'mb-genql'
+import { useSession } from 'next-auth/react'
 import { useParams } from 'next/navigation'
-import { useAIChat } from '@/lib/hooks/use-ai-chat'
-import { Message } from 'ai'
-import { Chatbot } from 'mb-genql/generated'
+import React, { useEffect } from 'react'
+import { toast } from 'react-hot-toast'
 import { useThread } from '@/lib/hooks/use-thread'
+import { botNames } from '@/lib/bots-names'
+import { useSidebar } from '@/lib/hooks/use-sidebar'
+import { useModel } from '@/lib/hooks/use-model'
+import { improveMessage } from '@/lib/helpers/ai-helpers'
+import { AiClientType } from '@/types/types'
 
 export function Chat({
   initialMessages,
@@ -21,58 +32,210 @@ export function Chat({
   scrollToBottom: scrollToBottomOfPopup,
   isAtBottom: isAtBottomOfPopup
 }: ChatProps) {
+  const { data: session } = useSession()
   const {
-    messages: allMessages,
-    append: appendWithMbContextPrompts,
-    reload,
-    stop,
-    isLoading,
-    input,
-    setInput,
-    isNewChat,
-  } = useAIChat(initialMessages ?? [], threadId, chatbot as Chatbot);
-  const { isOpenPopup, isAtBottom } = useThread();
+    allMessages: threadAllMessages,
+    initialMessages: threadInitialMessages,
+    activeThread,
+    setActiveThread,
+    setIsNewResponse,
+    setIsOpenPopup,
+    isOpenPopup,
+    sectionRef,
+    isAtBottom: isAtBottomOfSection
+  } = useThread()
+  const { activeChatbot } = useSidebar()
+  const containerRef = React.useRef<HTMLDivElement>()
 
-  const containerRef = useRef<HTMLDivElement>(null)
   const params = useParams<{ chatbot: string; threadId: string }>()
+  const isNewChat = Boolean(!params.threadId && !activeThread)
+  const { selectedModel, clientType } = useModel()
 
-  const sendMessageFromResponse = async (bulletContent: string) => {
-    const fullMessage = `Tell me more about ${bulletContent}`
-    appendWithMbContextPrompts({
-      role: 'user',
-      content: fullMessage
+  const { messages, append, reload, stop, isLoading, input, setInput } =
+    useChat({
+      initialMessages:
+        params.threadId || isNewChat
+          ? initialMessages?.filter(m => m.role === 'system')
+          : threadInitialMessages.filter(m => m.role === 'system'),
+      id: params.threadId || isNewChat ? threadId : activeThread?.threadId,
+      body: {
+        id: params.threadId || isNewChat ? threadId : activeThread?.threadId,
+        model: selectedModel,
+        clientType
+      },
+      onResponse(response) {
+        if (response.status === 401) {
+          toast.error(response.statusText)
+        }
+      },
+      onFinish(message) {
+        saveNewMessage({
+          role: 'assistant',
+          threadId:
+            params.threadId || isNewChat ? threadId : activeThread?.threadId,
+          content: message.content,
+          jwt: session!.user?.hasuraJwt
+        })
+      }
     })
-  }
 
-  // Debounced scroll to bottom
-  let timeoutId: NodeJS.Timeout
+  const { scrollY } = useScroll({
+    container: containerRef as React.RefObject<HTMLElement>
+  })
+
+  const { isAtBottom } = useAtBottom({
+    ref: containerRef,
+    scrollY
+  })
+
+  // ? saffer way to debounce scroll to bottom
+  let timeoutId: any
   const debounceScrollToBottom = (element: HTMLElement | undefined) => {
     clearTimeout(timeoutId)
     timeoutId = setTimeout(() => {
       scrollToBottomOfElement(element)
-    }, 150)
+      clearTimeout(timeoutId)
+    }, 150) //? Adjustable delay as necessary
   }
 
   const scrollToBottom = () => {
-    if (containerRef.current) {
-      debounceScrollToBottom(containerRef.current)
+    if (
+      (params.threadId && containerRef.current) ||
+      (!params.threadId && sectionRef.current)
+    ) {
+      let element: any
+      if (sectionRef.current) {
+        element = sectionRef.current
+      } else {
+        element = containerRef.current
+      }
+      debounceScrollToBottom(element)
     }
   }
+
+  // we merge past assistant and user messages for ui only
+  // we remove system prompts from ui
+  const allMessages =
+    params.threadId || isNewChat
+      ? uniqBy(initialMessages?.concat(messages), 'content').filter(
+          m => m.role !== 'system'
+        )
+      : uniqBy(threadAllMessages.concat(messages), 'content').filter(
+          m => m.role !== 'system'
+        )
+
+  const sendMessageFromResponse = async (bulletContent: string) => {
+    setIsNewResponse(true)
+    const fullMessage = `Tell me more about ${bulletContent}`
+    await saveNewMessage({
+      role: 'user',
+      threadId:
+        params.threadId || isNewChat ? threadId : activeThread?.threadId,
+      content: fullMessage,
+      jwt: session!.user?.hasuraJwt
+    })
+    append({
+      role: 'user',
+      content: `First, think about the following questions and requests: [${getAllUserMessagesAsStringArray(
+        allMessages
+      )}].  Then answer this question: ${fullMessage}`
+    })
+  }
+
+  // we extend append function to add our system prompts
+  const appendWithMbContextPrompts = async (
+    userMessage: Message | CreateMessage,
+    chatRequestOptions?: ChatRequestOptions
+  ) => {
+    let processedMessage = userMessage.content
+
+    console.log('Original message:', processedMessage)
+
+    try {
+      console.log('Original message:', userMessage.content)
+      processedMessage = await improveMessage(
+        userMessage.content,
+        clientType as AiClientType,
+        selectedModel
+      )
+      console.log('Refined message:', processedMessage)
+      if (processedMessage === userMessage.content) {
+        console.warn('Message was not improved by AI. Using original message.')
+      }
+    } catch (error) {
+      console.error('Error processing message:', error)
+      // Fall back to original message if processing fails
+      processedMessage = userMessage.content
+    }
+
+    if (isNewChat && chatbot) {
+      await createThread({
+        threadId,
+        chatbotId: chatbot.chatbotId,
+        jwt: session!.user?.hasuraJwt,
+        userId: session!.user.id,
+        isPublic: activeChatbot?.name !== 'BlankBot'
+      })
+      const thread = await getThread({
+        threadId,
+        jwt: session!.user?.hasuraJwt
+      })
+      setActiveThread(thread)
+      setIsOpenPopup(true)
+    }
+    if (activeThread?.threadId) {
+      setIsOpenPopup(true)
+    }
+
+    await saveNewMessage({
+      role: 'user',
+      threadId:
+        params.threadId || isNewChat ? threadId : activeThread?.threadId,
+      content: processedMessage,
+      jwt: session!.user?.hasuraJwt
+    })
+
+    setIsNewResponse(true)
+
+    return append(
+      isNewChat
+        ? { ...userMessage, content: processedMessage }
+        : {
+            ...userMessage,
+            content: `First, think about the following questions and requests: [${getAllUserMessagesAsStringArray(
+              allMessages
+            )}].  Then answer this question: ${processedMessage}`
+          }
+    )
+  }
+
+  useEffect(() => {
+    if (
+      params.chatbot &&
+      activeThread &&
+      botNames.get(params.chatbot) !== activeThread.chatbot.name
+    ) {
+      setIsOpenPopup(false)
+      setActiveThread(null)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
   useEffect(() => {
     if (isLoading && isOpenPopup && scrollToBottomOfPopup) {
       const timeout = setTimeout(() => {
         scrollToBottomOfPopup()
+        clearTimeout(timeout)
       }, 150)
-      return () => clearTimeout(timeout)
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isLoading, isOpenPopup, scrollToBottomOfPopup])
 
   return (
     <>
-      {params.threadId && (
+      {params.threadId ? (
         <div
-          ref={containerRef}
+          ref={containerRef as React.Ref<HTMLDivElement>}
           className={cn(
             'pb-[200px] pt-4 md:pt-10 h-full overflow-auto',
             className
@@ -84,17 +247,27 @@ export function Chat({
             sendMessageFn={sendMessageFromResponse}
           />
           <ChatScrollAnchor
-            isAtBottom={isPopup ? !!isAtBottomOfPopup : !!isAtBottom}
+            isAtBottom={
+              params.threadId
+                ? isAtBottom
+                : isPopup
+                  ? Boolean(isAtBottomOfPopup)
+                  : isAtBottomOfSection
+            }
             trackVisibility={isLoading}
           />
         </div>
-      )}
+      ) : null}
 
       {((isOpenPopup && isPopup) || (!isOpenPopup && !isPopup)) && (
         <ChatPanel
-          className={`${!chatbot ? 'hidden' : ''} ${chatPanelClassName}`}
-          scrollToBottom={isPopup && scrollToBottomOfPopup ? scrollToBottomOfPopup : scrollToBottom}
-          id={params.threadId || isNewChat ? threadId : undefined}
+          className={`${!activeThread && !activeChatbot ? 'hidden' : ''} ${chatPanelClassName}`}
+          scrollToBottom={
+            isOpenPopup && isPopup && scrollToBottomOfPopup
+              ? scrollToBottomOfPopup
+              : scrollToBottom
+          }
+          id={params.threadId || isNewChat ? threadId : activeThread?.threadId}
           isLoading={isLoading}
           stop={stop}
           append={appendWithMbContextPrompts}
@@ -111,7 +284,13 @@ export function Chat({
               : ''
           }
           showReload={!isNewChat}
-          isAtBottom={isPopup ? isAtBottomOfPopup : isAtBottom}
+          isAtBottom={
+            params.threadId
+              ? isAtBottom
+              : isPopup
+                ? Boolean(isAtBottomOfPopup)
+                : isAtBottomOfSection
+          }
         />
       )}
     </>
@@ -127,4 +306,12 @@ export interface ChatProps extends React.ComponentProps<'div'> {
   isPopup?: boolean
   scrollToBottom?: () => void
   isAtBottom?: boolean
+}
+
+export function getAllUserMessagesAsStringArray(allMessages: Message[]) {
+  const userMessages = allMessages.filter(m => m.role === 'user')
+  const cleanMessages = userMessages.map(m =>
+    extractBetweenMarkers(m.content, 'Then answer this question:')
+  )
+  return cleanMessages.join(', ')
 }
